Stage 1 : Basic separation and get and post

+++ src/PaymentGateway.Api/Models/Responses/PostPaymentResponse.cs
@@ -8,5 +8,7 @@ public class PostPaymentResponse
     public int ExpiryMonth { get; set; }
     public int ExpiryYear { get; set; }
     public string Currency { get; set; }
-    public int Amount { get; set; }
+    public decimal Amount { get; set; }
 }
+
+
